Refactoring + New Header
I wasn't confident in what I had begun to write as an input validation solution, so instead of pushing that, I'm pushing a change that just spaces out most of the .CODE program chunks, making them easier to read. I kept all of the labels the same, in the event that anything new would need to be written between now and Thursday regarding the main body. Other than adding more prompts to instruct/inform the user of what to do (I have some template strings I used in my Program #1 that I've included here too, unused), I/O getting another loop stapled on seems to be all the program needs.

A lot of the stack related functions feel like they could be wrapped into for loops; if that sort of code compacting sounds useful. Since I didn't share my attempts at the I/O mentioned, I'll just briefly elaborate here as to what the pseudo code was:

- compare the first entry of choice with 0, jump to a loop initializer if it is not equal,
- enter a loop that terminates when either a counter == len, or, an ascii character that is not TAB or SPACE is found
   + if nothing is found, this jumps back to start (and presumably resets some registers in the process; unless the resetting in start is moved to the top instead of the bottom of the loop),
   + if something is found that is a digit, preserve the register being used as an iterator for another loop below,
   + if something is found that is not a digit, enter the op section of MAIN using the counter's offset instead of choice[0],
- start another loop for digits, which iterates until the current counter register == len, or, a non-digit character is found.
   + every digit character is compared against 0-9, with jne's to a loop terminator at each comparison.
   + for the actual conversion, https://stackoverflow.com/questions/19461476/convert-string-to-int-x86-32-bit-assembler-using-nasm seems to offer a decent (and compact) solution. Ideally the string conversion in Irvine's library would be used with a new concatenated string instead, but that option is currently off the table.
- anything after an operator/+-number is ignored in an input line, though wrapping this entire process into another loop that saves where the process terminates to be copied to a counter could allow for macros (1+1V, or something similar).

I'll definitely try to start on some of the comparison logic within the next couple of days, though Java Program #3 also is due next week, and it may be a safer bet to work on that first, even if it seems comparatively similar. I'm mostly sharing the pseudocode to be collaborative in the creation process, not so much to leave it as instructions for something to be written, though hopefully it's legible either way.