Mostly Implemented (Not Working)
I have written a mostly working iterator that goes through program input one step at a time, and accounts for most of the operator(s). The main issue I'm having is with the register/memory indexing, since it seems like the loop is entering an unintended success/fail state where it never prints anything. I couldn't get the debugging controls in Visual Studio to be of any help, though this is likely an off-by-one sort of problem, somewhere.

Something else to note which I didn't get to yet; indexing for the stack is meant to begin at 0 and end at 28, with the counter incrementing by four (starts at -4). This just involves swapping the numbers out, which is why it's lower priority than new I/O code. 

If we can't get this/an equivalent of this code working, I'm fine with trying a different solution with the main file (hence why this is a branch), or just polishing up what we have and submitting that, with some easier to implement changes.