New vars to get the info entered
I decided to put the info the user enters in their own variables. So the command has its own var, operand 1 has its own variable, and so on with the other 2 operands. If the user doesn't enter anything, the vars are set to null, with the exception of operand 2. This is due to that var having 0 as a valid value, so I initialize it to -1 when info isn't entered. 

A var is also set to null if it is out of the valid range(0-7 for priority, and 1-50 for run_time).

As of now, the program is just displaying all the information entered, skipping whitespace of course. The rest should be very easy as we have all the info we need and basically just need to make the records now.