Added part of the kill procedure...SHOW proc needs fixed
The show PROC needs to be fixed because it will make the kill procedure much easier. The show PROC should print the jobs by checking the status and print everything that' does not have a status of 0(available). Once thats fixed the kill procedure just need to put null in the 14 bytes that are pointed to by jobnamepos