Final commit, with known issues
There are a couple of problems with this implementation of the program that I cannot fix. The associated documentation alongside the program rubric says that each command should have verbose output lines, depending on what happens, and there just wasn't enough time, even with me working all night to do that. Another thing is the non-standard (see the calculator program, please) digit validation for operators is really prone to errors when something too big is entered. Input prompts also do not loop if they get bad input. Otherwise, STEP works as intended, and I did my best to pretty up what I could, but most of my time was spent making the program file usable, and diagnosing issues in LOAD/others. I expect a decent grade still, but, this was not fun to fix.