Code Skeleton (without nodes)
The current progress I've made mostly is refining the command handler, and associated functions. Since file opening/writing code already was provided via the discord in a text file, this means the only actual point of contention is the node code (which Josh said he was working on, to some degree). Not all procedures in this revision are transparent yet, and some have incomplete/sparse comments, though the codebase should be much easier to work with overall. I also altered how commands are copied to be non-destructive to the input buffer in a procedure, which I suggest is reused with a larger buffer for copying the file name. File stepping requires building strings, which is really similar to fetching commands, except set strings are copied before and after the text in question.